from_state_vector: Fixed arg_pe incorrect when it should be > 180°.
Adds a missing check for the case where the eccentricity vector is in the third
or fourth quadrant, requiring arg_pe to be negated.

Note that this causes an edge case where arg_pe should be 0° but it comes out as
360° because we are not using modulo, but rather adding 2*pi. I didn't want to
fix that in this commit because all the calculations in from_state_vector behave
the same way.

Fixes #39.