from_state_vector: Fixed assertion error for retrograde non-inclined orbits.
This was calculating both raan and f as nan, since i=180° was going into the
inclined orbit code path, when it is actually non-inclined. When invoked via
KeplerianElements.from_state_vector, the nan value would fail an assertion.

This also adds an important case needed to get the correct value of arg_pe
for retrograde non-inclined orbits: if i=180°, arg_pe needs to be inverted since
the orbit itself is upside down. Added two regression test cases for this case.

Fixes #18 and #25.