from_state_vector: Fixed arg_pe bug where orbit is very slightly inclined.
If the Z coordinate is negative, but the inclination is extremely small, it
triggers the non-inclined case, which hard-codes raan to 0, yet still runs the
code path which inverts arg_pe if ev.z < 0. This code path only makes sense if
raan is also rotated 180°, which it is not.

Moved the ev.z < 0 code path into the i-not-small condition. This also makes the
code more readable, as now all the code that sets arg_pe is in one place.

Added regression test cases for this issue.

Fixes #44.